[BE] feat: photo 서비스 구현 #95
+ 이미지 로딩, 인코딩 유틸 함수 담당
